Move flagbuilder to a shared package
Splitting up the CoreOS PR into manageable chunks
