+++ pkg/flagbuilder/build_flags.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package nodeup
+package flagbuilder
 
 import (
 	"fmt"
@@ -25,8 +25,8 @@ import (
 	"strings"
 )
 
-// buildFlags is a template helper, which builds a string containing the flags to be passed to a command
-func buildFlags(options interface{}) (string, error) {
+// BuildFlags builds flag arguments based on "flag" tags on the structure
+func BuildFlags(options interface{}) (string, error) {
 	var flags []string
 
 	walker := func(path string, field *reflect.StructField, val reflect.Value) error {
